Make genfut a library and use that in binary.
