+++ Cargo.toml
@@ -7,11 +7,13 @@ edition = "2018"
 [dependencies]
 bindgen = "0.51.0"
 cc = "1.0.45"
+Inflector = "0.11.4"
 regex = "1"
 structopt = "0.3.2"
 
+
 [features]
-default = ["sequential_c"]
 sequential_c = []
 cuda = []
 opencl = []
+no-futhark = []
