+++ src/genc.rs
@@ -3,6 +3,9 @@ use std::io::{self, Write};
 use std::path::PathBuf;
 use std::process::Command;
 
+#[cfg(feature = "no-futhark")]
+pub(crate) fn gen_c(in_file: &std::path::Path, out_dir: &std::path::Path) {}
+
 #[cfg(feature = "sequential_c")]
 pub(crate) fn gen_c(in_file: &std::path::Path, out_dir: &std::path::Path) {
     let out_path = PathBuf::from(out_dir);
